requirements.txt
numpy
numba
matplotlib
cupy-cuda12x
